Fix & refactor
